+++ base-compat/src/Debug/Trace/Compat.hs
@@ -5,6 +5,11 @@ module Debug.Trace.Compat (
 , traceShowId
 , traceM
 , traceShowM
+, traceWith
+, traceShowWith
+#if MIN_VERSION_base(4,5,0)
+, traceEventWith
+#endif
 ) where
 
 #if !(MIN_VERSION_base(4,7,0)) || MIN_VERSION_base(4,9,0)
@@ -16,7 +21,7 @@ import Debug.Trace as Base hiding (
   )
 #endif
 
-#if !(MIN_VERSION_base(4,9,0))
+#if !(MIN_VERSION_base(4,18,0))
 import Prelude.Compat
 #endif
 
@@ -74,4 +79,40 @@ Like 'traceM', but uses 'show' on the argument to convert it to a 'String'.
 -}
 traceShowM :: (Show a, Applicative f) => a -> f ()
 traceShowM = traceM . show
+#endif
+
+#if !(MIN_VERSION_base(4,18,0))
+{-|
+Like 'trace', but outputs the result of calling a function on the argument.
+
+>>> traceWith fst ("hello","world")
+hello
+("hello","world")
+
+/Since: 4.18.0.0/
+-}
+traceWith :: (a -> String) -> a -> a
+traceWith f a = trace (f a) a
+
+{-|
+Like 'traceWith', but uses 'show' on the result of the function to convert it to
+a 'String'.
+
+>>> traceShowWith length [1,2,3]
+3
+[1,2,3]
+
+/Since: 4.18.0.0/
+-}
+traceShowWith :: Show b => (a -> b) -> a -> a
+traceShowWith f = traceWith (show . f)
+
+# if MIN_VERSION_base(4,5,0)
+-- | Like 'traceEvent', but emits the result of calling a function on its
+-- argument.
+--
+-- /Since: 4.18.0.0/
+traceEventWith :: (a -> String) -> a -> a
+traceEventWith f a = traceEvent (f a) a
+# endif
 #endif